// src/models/buku.dto.ts
export interface bukuCreate {
  judul: string;
  penulis: string;
  kategori: string;
  penerbit: string;
  jumlahStok: number;
  tahunPenerbit: number;
}

export interface bukuUpdated {
  judul: string;
  penulis: string;
  kategori: string;
  penerbit: string;
  jumlahStok: number;
  tahunPenerbit: number;
}

export interface bukuDelete {
  judul: string;
  penulis: string;
  kategori: string;
  penerbit: string;
  jumlahStok: number;
  tahunPenerbit: number;
}